requirements.txt
langgraph
openai
streamlit
yfinance
dotenv
tavily-python